Update 20240506
Update two dataloader, 'bbox_reader_neg.py' load nodule with dicom
based method and random choose a negative samples to each dicom.
'bbox_reader_beg_nodulebased.py' load nodule with nodule based
method like original one. and then randomly choose to use positive
nodule or use a random negative sample while loading each GT nodule.

Add new collate method for dicom based loading method.

Other bug fixed:
    'pybox.py/torch_nms': check data type which has to be float32
    'rpn_loss.py/weighted_focal_loss_with_logits_OHEM':
        force to pick hard neg samples for training even pos=0
    'rpn_loss.py/rpn_loss': return reg_losses correctly
    'util.py/crop_with_lobe': check align data type and value range

+++ single_test.py
@@ -64,7 +64,7 @@ def main():
         test_name = args.test_name
         data_dir = args.data_dir
         annotation_dir = args.annotation_dir
-        label_types = config['label_types'][0]
+        label_types = config['label_types']
         augtype = args.augtype
         # num_workers = config['num_workers']
         num_workers = 1
@@ -91,7 +91,7 @@ def main():
                               num_workers=num_workers, pin_memory=False, collate_fn=train_collate)
         # breakpoint()
         print('out_dir', out_dir)
-        save_dir = os.path.join(out_dir, 'res', str(epoch)+'_')
+        save_dir = os.path.join(out_dir, 'res', str(epoch)+'')
         if not os.path.exists(save_dir):
             os.makedirs(save_dir)
             
@@ -196,9 +196,9 @@ def eval(net, dataset, annotation_dir, data_dir, save_dir=None, create_data=True
     rpn_submission_path = os.path.join(eval_dir, 'submission_rpn.csv')
     # rcnn_submission_path = os.path.join(eval_dir, 'submission_rcnn.csv')
     # ensemble_submission_path = os.path.join(eval_dir, 'submission_ensemble.csv')
-
-    df = pd.DataFrame(rpn_res, columns=col_names)
-    df.to_csv(rpn_submission_path, index=False)
+    if not os.path.exists(rpn_submission_path):
+        df = pd.DataFrame(rpn_res, columns=col_names)
+        df.to_csv(rpn_submission_path, index=False)
 
     # df = pd.DataFrame(rcnn_res, columns=col_names)
     # df.to_csv(rcnn_submission_path, index=False)
